Add Ticker datasets to neon.
This commit adds a Ticker class in the Neon data directory.

This class is meant to be used for synthetic algorithmic tasks
such as those seen in the Neural Turing Machines paper or in
the Neural GPUs paper.

The commit also adds tasks from the NTM paper, including
copy, repeat copy, etc.

Finally, it adds a callback for inspecting input target tuples
for Ticker datasets next to the model's output during training.
This can be useful for debugging, and is also fun to watch.