Make batch writing implicit
Macrobatches are now written out as part of data loading. There's no
need to run the batch writer as a pre-requisite to training. As soon
as the first macrobatch file is available, training can commence (note
that this is only true if there's no metadata to be aggregated over
the entire dataset).