Fix dependency issue in loader makefile
Also, let the loader library build even when OpenCV libraries are absent.
A warning is printed during the build. In this case, any attempt to use the
loader library results in a runtime error and a message asking the user to
install the pre-requisites.

The build output of the loader library is no longer hidden in the top-level
makefile.