Fixes I1K batch_writer data structures
I removed the generic transformation code in the `Ingester` base class because
it assumed access to filenames which is not the case now with Imagenet (which
was the only child class to use this transformation capability).

I tried using multiprocessing to convert images in parallel, but found no
speedup: imagemagick is multithreaded itself, and the IPC costs outweigh the
extra CPU usage, and it breaks the nice nested progress bars that I added.

Bring imagenet ingest down to 1.5 hours

Fix imagenet ingest docs

Switch to uncompressed manifest files for dataloader