Fixes python 3 builds on Ubuntu 16.04
Newer systems often don't have a `python` symlink to python3 which causes the
variable expansion leading up to $(ACTIVATE) to fail with the message:

Makefile:100: *** target pattern contains no '%'.  Stop.

This works around that problem in a minimal fashion. A more correct fix would
be to better detect versions of Python, but that is out of scope for the moment.